type data uint8 to int64

+++ app/models/akun.go
@@ -5,7 +5,7 @@ import (
 )
 
 type Akun struct {
-	IDUser   uint8  `json:"id_user" gorm:"primary_key" column:"id_user"`
+	IDUser   int64  `json:"id_user" gorm:"primary_key" column:"id_user"`
 	Email    string `gorm:"default:not null" column:"email"`
 	Password string `gorm:"default:not null" column:"password"`
 	orm.Timestamps
