+++ app/models/jabatan.go
@@ -5,7 +5,7 @@ import (
 )
 
 type Jabatan struct {
-	IdJabatan uint8  `json:"id_jabatan" gorm:"primary_key"`
+	IdJabatan int64  `json:"id_jabatan" gorm:"primary_key"`
 	Name      string `json:"name" gorm:"default:not null" column:"name"`
 	orm.Timestamps
 }
